Updated CI setup
