+++ .travis.yml
@@ -1,4 +1,3 @@
-language: python
-python: '3.7'
+language: minimal
 services: docker
-script: docker run --rm -tv $(pwd):/src -w /src ghdl/vunit:gcc-master python3 /src/run.py
+script: docker run --rm -tv $(pwd):/src -w /src ghdl/vunit:llvm-master python3 /src/run.py
